refactor: 타입스크립트 types.ts -> types.d.ts로 변경 - namespace로 관리
