+++ apps/admin/app/contents/[testId]/page.tsx
@@ -1,7 +1,5 @@
 import ContenDetalis from '@/containers/ContentDetalis';
 
 export default function Page() {
-  return (
-      <ContenDetalis />
-  );
+  return <ContenDetalis />;
 }
