+++ apps/admin/app/page.tsx
@@ -1,4 +1,4 @@
-import Dashboard from "@/containers/Dashboard";
+import Dashboard from '@/containers/Dashboard';
 
 export default function Page() {
   return <Dashboard />;
