+++ apps/admin/utils/getSelectOptions.ts
@@ -1,4 +1,3 @@
-
 export type SelectOptions = {
   value: string;
   label: string;
