+++ apps/client/components/base/MbtiTestContent.tsx
@@ -1,14 +1,12 @@
 import { FONT, CONST_MAIN_PAGE } from '@/constants/constant';
 import { MbtiTestcommentCountImage, MbtiTestLikeCountImage, MbtiTestPlayCountImage } from '@/public/images/mbtiTest';
 
-import { Wrap_mediaquery } from '@/components/ui/wrap/Wrap';
-import { MbtiTestVersionBig, MbtiTestVersionSmallForSeveral } from '@/components/ui/mbtiTest';
+import { Wrap_mediaquery } from '@/components/ui/Wrap';
+import { MbtiTestVersionBig, MbtiTestVersionSmallForSeveral } from '@/components/ui/MbtiTest';
 import { Image, Text, Wrap } from '@/components/ui/CommonElements';
 import { TitleText } from '@/components/ui/CommonElements';
-import * as TypesMbtiTestContent from '@/components/ui/types/mbtiTestContent';
-import * as TypesBase from '@/components/base/types';
 
-export function TitleAndText({ text }: TypesMbtiTestContent.TitleAndTextProps) {
+export function TitleAndText({ text }: Base.TitleAndTextProps) {
   return (
     <Wrap_mediaquery flexDirection="column" justifyContent="space-around" alignItems="baseline">
       <TitleText fontSize={FONT.SIZE.EXTRA_LARGE} fontWeight={FONT.BOLD_SCALE.BOLD}>
@@ -21,7 +19,7 @@ export function TitleAndText({ text }: TypesMbtiTestContent.TitleAndTextProps) {
   );
 }
 
-export function TitleAndMbtiTestBig({ detail }: TypesMbtiTestContent.TitleAndMbtiTestProps) {
+export function TitleAndMbtiTestBig({ detail }: Base.TitleAndMbtiTestProps) {
   return (
     <Wrap_mediaquery
       flexDirection="column"
@@ -38,9 +36,7 @@ export function TitleAndMbtiTestBig({ detail }: TypesMbtiTestContent.TitleAndMbt
   );
 }
 
-export function TitleAndMbtiTestsSmallForSeveral({
-  mbtiTestData,
-}: TypesMbtiTestContent.TitleAndMbtiTestsSmallForSeveralProp) {
+export function TitleAndMbtiTestsSmallForSeveral({ mbtiTestData }: Base.TitleAndMbtiTestsSmallForSeveralProp) {
   return (
     <Wrap_mediaquery
       flexDirection="column"
@@ -57,7 +53,7 @@ export function TitleAndMbtiTestsSmallForSeveral({
   );
 }
 
-export function MbtiTestCountImageArea({ countData }: TypesBase.MbtiTestCountImageAreaProp) {
+export function MbtiTestCountImageArea({ countData }: Base.MbtiTestCountImageAreaProp) {
   return (
     <Wrap position="relative" top="-2.5rem" display="flex">
       <Wrap display="flex" margin="0 0.7rem 0 0 ">
