+++ apps/client/components/base/MyFooter.tsx
@@ -5,7 +5,7 @@ import { usePathname } from 'next/navigation';
 
 import { FONT, CONST_FOOTER, IMAGE_ALT_STRING } from '@/constants/constant';
 
-import { Wrap_mediaquery } from '@/components/ui/wrap/Wrap';
+import { Wrap_mediaquery } from '@/components/ui/Wrap';
 import { Wrap, Text, Image } from '@/components/ui/CommonElements';
 import { PolicyText, CopyrightText } from '@/components/styledComponents';
