--- apps/client/components/ui/types/square.ts
@@ -1,3 +1,0 @@
-export type MbtiTestTitleBlackSquareDivProp = {
-  bottom: string;
-};
