--- apps/client/components/ui/types/wrap.ts
@@ -1,11 +1,0 @@
-export type WrapMediaqueryProp = {
-  width?: string;
-  margin?: string;
-  flexDirection?: string;
-  justifyContent?: string;
-  alignItems?: string;
-  padding?: string;
-  position?: string;
-  textAlign?: string;
-  flexWrap?: string;
-};
