+++ apps/client/containers/viewLatestMbtiTest/index.tsx
@@ -4,8 +4,8 @@ import * as Types from '@/containers/types/viewLatestMbtiTest';
 import { VIEW_MBTI_TEST_PAGE } from '@/constants/constant';
 
 import { TitleAndText } from '@/components/base/MbtiTestContent';
-import { Wrap_mediaquery } from '@/components/ui/wrap/Wrap';
-import { MbtiTestForViewPage } from '@/components/ui/mbtiTest';
+import { Wrap_mediaquery } from '@/components/ui/Wrap';
+import { MbtiTestForViewPage } from '@/components/ui/MbtiTest';
 
 const text = {
   titleText: VIEW_MBTI_TEST_PAGE.LATEST.TITLE_TEXT,
