+++ apps/client/hooks/types.d.ts
@@ -1,0 +1,7 @@
+declare namespace Hooks {
+  type containerRefCurrent = {
+    current: HTMLDivElement | null;
+  };
+
+  type animationData = any;
+}
