--- apps/client/hooks/types.ts
@@ -1,5 +1,0 @@
-export type containerRefCurrent = {
-  current: HTMLDivElement | null;
-};
-
-export type animationData = any;
