+++ apps/client/services/types.d.ts
@@ -1,0 +1,9 @@
+declare namespace Services {
+  type FetchClientProp = {
+    url: string;
+    method: string;
+    headers?: {
+      Authorization: string | null;
+    };
+  };
+}
