--- apps/client/services/types.ts
@@ -1,7 +1,0 @@
-export type FetchClientProp = {
-  url: string;
-  method: string;
-  headers?: {
-    Authorization: string | null;
-  };
-};
